pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rpgxp"
version = "0.0.1"
authors = [
    { name="Andrew Foote", email="footeandrew1@gmail.com" },
]
description = "Processes data from games made with RPG Maker XP."
readme = "README.md"
requires-python = ">=3.14.0a1"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
    "Development Status :: 1 - Planning",
    "Typing :: Typed",
]
dependencies = [
    "apsw>=3.46.1.0",
    "numpy>=2.1.2",
    "jinja2>=3.1.4",
    "sortedcontainers>=2.4.0",
    "pillow",
    "ruby-marshal-parser",
]

[tool.uv.sources]
ruby_marshal_parser = { path = "../ruby-marshal-parser", editable = true }

[project.urls]
Homepage = "https://github.com/Andrew-Foote/rpgxp"
Issues = "https://github.com/Andrew-Foote/rpgxp/issues"

[project.optional-dependencies]
dev = [
    "mypy >= 1.13.0",
    "sortedcontainers-stubs>=2.4.2",
    "pytest>=8.3.3",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.mypy]
packages = ["rpgxp", "tests", "golden"]
